Allow users to define the aspect ratio of prov report images
Add a new optional parameter `aspect_ratio` to the `ProvReportView`
This is used to set the aspect ratio of the svg and jpg images produced
for the prov report. The value of `aspect_ratio` should be a float. If
is not a float or not present then a value of 0.71 will be used. 0.71 is
the aspect ratio of landscape A4 paper