Changes to the way RO Crate handles external objects
External files may point directly to data, in which case they will be
used directly as
inputs to a `CodeRun`. External files will have a link to them in the
metadata file, but
will not be packaged in the zip file. However, it maybe that data has
had to be
extracted from an external file before it can be used by a `CodeRun`,
i.e. from a
journal article, In which case there will be an associated `DataProduct`
that would have
been made to contain the data so that it can be used in a `CodeRun`. If
this is the case
the relationship between the external file and `DataProduct` is modelled
as a RO Crate
ContextEntity` of type `CreateAction`.